:zap: Predicted forms: B/D/F/G of Ch. 1 (#12)
- Add predicted forms for sections B, D, F, and G
Information for sections B, D, and G were largely based on @ysm051222's
research, though I have supplanted certain missing words. Information
for section F was gathered by me.

In section B, the predicted form box for '마루', which was given by
@ysm051222 as '말\~물\~몰+리\~레, 반침', was left empty. This is because I
strongly suspect that 말리, 말레, 물리, 물레, 몰리, 몰레 refer to 마루 as in the top
of a mountain or roof (from Middle Korean ᄆᆞᄅᆞ~ᄆᆞᆯㄹ < pre-Middle Korean
*ᄆᆞᄅᆞᆯ), not 마루 as in the space in a traditional Korean house (from
Middle Korean 마로). I expect that the true dialectal form of this word
would just be 마루, but I could not find it in any sources I referenced,
so I left it empty for now.

Any suggestions and/or additional fixes are welcome.